+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-Nothing Here Yet
+### Added
+- Color Picker
 
 ## [0.1.6] - 2019-03-08